Separate defines for this library from those of deps
This makes it easier to write this list to the config-mode module and
let it merge with those from dependencies in the user's program
(although that implies that we no longer pretend that inconsistencies
are not possible).

The convention that singular form contains values for this project
alone whereas plural form contains values for both this project and
dependencies, is used.